requirements.txt
requests>=2.31.0
pyTelegramBotAPI>=4.14.0
beautifulsoup4>=4.12.2
openai>=1.51.0
httpx>=0.27.2
anyio>=4.4.0
PyPDF2>=3.0.1
python-dotenv>=1.0.0
google-genai>=1.0.0
opencv-python>=4.8.0
